Removed Duplicate Gamma function
before Cyrus' commit, there was a Gamma function placeholder, after Cyrus' commit, the placeholder was still there so it was creating a conflict (2 functions named the same), so I just removed it from master (even though I shouldn't have done it that way, its easier to do it like that). I also updated the Readme to not have the generic text we added.